wip refactor aim functions
