+++ malariagen_data/anoph/aim_params.py
@@ -1,0 +1,8 @@
+from typing_extensions import Annotated, TypeAlias
+
+aims: TypeAlias = Annotated[
+    str,
+    """
+    Which set of ancestry informative markers to use.
+    """,
+]
